Use abstract-blob-store typings
And remove mostly unused async blob store wrapper
